UPet-api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
